page jumps and reorganize favorites
