+++ products/templates/items/products_by_owner.html
@@ -20,8 +20,8 @@
                 <div class="store-details">
                     Category
                     <a href="{% url 'products:items:category' item.category.id %}">{{item.category.name}}</a><br>
-                    <a href="{% url 'orders:request_donation' item.id %}">Request donation</a><br>
-                    <a href="{% url 'products:items:add_to_favorites_item_view' item.id %}">Add to favorites</a>
+                    <a href="{% url 'orders:item_view_request_donation' item.id %}">Request donation</a><br>
+                    <a href="{% url 'products:favorites:add_to_favorites_item_view' item.id %}">Add to favorites</a>
                     {% if item.status == 1 %}
                         <p>No requests yet for this product </p>
                     {% else %}
