+++ products/views/items.py
@@ -1,9 +1,7 @@
-from django.shortcuts import redirect, reverse, render, Http404
+from django.shortcuts import render, Http404
 from django.views.generic import ListView
 from products.models import Products
-from utils.favorites import Favorites
 from django.core.paginator import Paginator
-from django.urls import resolve
 from products.forms.filter import SearchItems
 
 # class ItemsList(ListView):
@@ -16,62 +14,17 @@ from products.forms.filter import SearchItems
 def list_view(request):
     form = SearchItems(request.GET)
     products = form.get_filtered_items()
-    # if request.user.is_authenticated:
-    #     products = Products.objects.exclude(status=3).exclude(user_id=request.user.id)
-    # else:
-    #     products = Products.objects.exclude(status=3)  # 3=inactive (see constants)
-
+    # products = products.filter(status__lt=3).exclude(user_id=request.user.id)
+    products = products.filter(status__lt=3)
     paginator = Paginator(products, 9)
     page_number = request.GET.get('page', 1)
     page_obj = paginator.get_page(page_number)
-    # form = SearchItems(request.GET)
-    # (print('request.GET', request.GET))
 
     return render(request, 'items/list.html', {
         'page_obj': page_obj,
         'items_list': products,
-        'form': form
+        # 'form': form
     })
-
-
-def add_to_favorites(request, item_id, page_num=None):
-    # next_url = request.GET.get('next')
-    favorites = Favorites(request.user, request.session)
-    favorites.add(item_id)
-
-    current_url = resolve(request.path_info).url_name
-    previous_url = request.META['HTTP_REFERER']
-    # print('request.GET', request.GET)
-    # print('request.path', request.path)
-    # print('request.meta', request.META['HTTP_REFERER'])
-
-    if current_url == 'add_to_favorites_item_view':
-        return redirect(reverse('products:items:item', args=(item_id,)))
-
-    if 'search_term' in previous_url:
-        return redirect(previous_url)
-
-    return redirect('%s?page=%s' % (
-        reverse('products:items:list'),
-        page_num
-
-    ))
-
-
-# def add_to_favorites(request, item_id, page_num=None):
-#     # next_url = request.GET.get('next')
-#     favorites = Favorites(request.user, request.session)
-#     favorites.add(item_id)
-#
-#     current_url = resolve(request.path_info).url_name
-#
-#     if current_url == 'add_to_favorites_item_view':
-#         return redirect(reverse('products:items:item', args=(item_id,)))
-#
-#     return redirect('%s?page=%s' % (
-#         reverse('products:items:list'),
-#         page_num
-#     ))
 
 
 def item_view(request, item_id):
